Broke condition.tsx into seperate files
