+++ web/src/designer/components/condition/types.tsx
@@ -1,0 +1,27 @@
+// Copyright 2023 FAIMS Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+export type ConditionType = {
+  operator: string;
+  field?: string;
+  value?: unknown;
+  conditions?: ConditionType[];
+};
+
+export type ConditionProps = {
+  onChange?: (v: ConditionType | null) => void;
+  initial?: ConditionType | null;
+  field?: string; // the field this condition will attach to
+  view?: string; // the view this condition will attach to
+};
